LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/mo-lspace-rspace-5-expected.txt


FAIL lspace/rspace spacing correctly set around operator (child of mrow, not embellished, LTR, normal text). assert_approx_equals: expected 25 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, not embellished, LTR, largeop variant). assert_approx_equals: expected 75 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, not embellished, LTR, stretchy variant). assert_approx_equals: expected 125 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, not embellished, LTR, glyph assembly). assert_approx_equals: expected 175 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, not embellished, LTR, normal text). assert_approx_equals: expected 25 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, not embellished, LTR, largeop variant). assert_approx_equals: expected 75 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, not embellished, LTR, stretchy variant). assert_approx_equals: expected 125 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, not embellished, LTR, glyph assembly). assert_approx_equals: expected 175 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, not embellished, LTR, normal text). assert_approx_equals: expected 25 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, not embellished, LTR, largeop variant). assert_approx_equals: expected 75 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, not embellished, LTR, stretchy variant). assert_approx_equals: expected 125 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, not embellished, LTR, glyph assembly). assert_approx_equals: expected 175 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, embellished, LTR, normal text). assert_approx_equals: expected 25 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, embellished, LTR, largeop variant). assert_approx_equals: expected 75 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, embellished, LTR, stretchy variant). assert_approx_equals: expected 125 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, embellished, LTR, glyph assembly). assert_approx_equals: expected 175 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, embellished, LTR, normal text). assert_approx_equals: expected 25 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, embellished, LTR, largeop variant). assert_approx_equals: expected 75 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, embellished, LTR, stretchy variant). assert_approx_equals: expected 125 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, embellished, LTR, glyph assembly). assert_approx_equals: expected 175 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, embellished, LTR, normal text). assert_approx_equals: expected 25 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, embellished, LTR, largeop variant). assert_approx_equals: expected 75 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, embellished, LTR, stretchy variant). assert_approx_equals: expected 125 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, embellished, LTR, glyph assembly). assert_approx_equals: expected 175 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, not embellished, RTL, normal text). assert_approx_equals: expected 50 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, not embellished, RTL, largeop variant). assert_approx_equals: expected 100 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, not embellished, RTL, stretchy variant). assert_approx_equals: expected 150 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, not embellished, RTL, glyph assembly). assert_approx_equals: expected 200 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, not embellished, RTL, normal text). assert_approx_equals: expected 50 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, not embellished, RTL, largeop variant). assert_approx_equals: expected 100 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, not embellished, RTL, stretchy variant). assert_approx_equals: expected 150 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, not embellished, RTL, glyph assembly). assert_approx_equals: expected 200 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, not embellished, RTL, normal text). assert_approx_equals: expected 50 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, not embellished, RTL, largeop variant). assert_approx_equals: expected 100 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, not embellished, RTL, stretchy variant). assert_approx_equals: expected 150 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, not embellished, RTL, glyph assembly). assert_approx_equals: expected 200 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, embellished, RTL, normal text). assert_approx_equals: expected 50 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, embellished, RTL, largeop variant). assert_approx_equals: expected 100 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, embellished, RTL, stretchy variant). assert_approx_equals: expected 150 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mrow, embellished, RTL, glyph assembly). assert_approx_equals: expected 200 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, embellished, RTL, normal text). assert_approx_equals: expected 50 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, embellished, RTL, largeop variant). assert_approx_equals: expected 100 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, embellished, RTL, stretchy variant). assert_approx_equals: expected 150 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of math, embellished, RTL, glyph assembly). assert_approx_equals: expected 200 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, embellished, RTL, normal text). assert_approx_equals: expected 50 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, embellished, RTL, largeop variant). assert_approx_equals: expected 100 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, embellished, RTL, stretchy variant). assert_approx_equals: expected 150 +/- 1 but got 0
FAIL lspace/rspace spacing correctly set around operator (child of mtd, embellished, RTL, glyph assembly). assert_approx_equals: expected 200 +/- 1 but got 0
22
+++
4444
⫿
55555
⥜
666666
⥜
7777777
22
+++
4444
⫿
55555
⥜
666666
⥜
7777777
22+++4444⫿55555⥜666666⥜7777777
22
+++
4444
⫿
55555
⥜
666666
⥜
7777777
22
+++
4444
⫿
55555
⥜
666666
⥜
7777777
22
+++
4444
⫿
55555
⥜
666666
⥜
7777777
22
+++
4444
⫿
55555
⥜
666666
⥜
7777777
22
+++
4444
⫿
55555
⥜
666666
⥜
7777777
22+++4444⫿55555⥜666666⥜7777777
22
+++
4444
⫿
55555
⥜
666666
⥜
7777777
22
+++
4444
⫿
55555
⥜
666666
⥜
7777777
22
+++
4444
⫿
55555
⥜
666666
⥜
7777777
